feat: Split EnemyStateMachine into separated states
