+++ Assets/Enemy/Prefabs/Fantasy Wolf/Materials/M_FantasyWolf_Legacy.mat
@@ -115,7 +115,7 @@ Material:
     - _ZWrite: 1
     m_Colors:
     - _BaseColor: {r: 1, g: 1, b: 1, a: 0.066}
-    - _Color: {r: 1, g: 1, b: 1, a: 1}
+    - _Color: {r: 1, g: 1, b: 1, a: 0.066}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.9411765, g: 0.9411765, b: 0.9411765, a: 1}
   m_BuildTextureStacks: []
