+++ Assets/Enemy/Scripts/States/EnemyStateMachine_Attack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ccd5c1ba454bcf94cb12b4e9b8a5baa8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
