+++ Assets/Enemy/Scripts/States/EnemyStateMachine_Chase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 66cfac7e4d04ebf49846aa2caf7592fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
