+++ Assets/Enemy/Scripts/States/EnemyStateMachine_Dead.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e59cddf89bca6f54b82be2a17a87a7d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
