+++ Assets/Enemy/Scripts/States/EnemyStateMachine_Idle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29a019fc3bbbc39488d8fcb08005bc66
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
