+++ Assets/Enemy/Scripts/States/EnemyStateMachine_Patrolling.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fe3568b7741f452489f00ed3adeccd36
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
