+++ Assets/Enemy/Scripts/States/EnemyStateMachine_Retreat.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public partial class EnemyStateMachine : MonoBehaviour
+{
+	private void RetreatState()
+	{
+
+	}
+}
