+++ Assets/Enemy/Scripts/States/EnemyStateMachine_Warning.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 234bb1e33bf32374eb3a7ebd9d7b4dc7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
